pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "qbo-pipeline"
version = "0.1.0"
description = "Extract QuickBooks data from n8n webhook and load into Supabase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<1",
  "python-dotenv>=1,<2",
  "psycopg2-binary>=2.9,<3",
  "google-genai>=1,<2",
  "openai>=1.40,<2",
  "sqlglot>=26,<27",
  "flask>=3,<4",
  "flask-cors>=4,<6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
